Removed user error print; messages updated
The urosUserErrPrintf() was clueless, and was removed.
Error message parameters are now ANSI C compliant, even if the GNU version was
more expressive.

Andrea Zoppi

+++ src/urosBase.c
@@ -173,7 +173,8 @@ void *urosAlloc(size_t size) {
 #if UROS_BASE_C_USE_ERROR_MSG
   void *chunk = uros_lld_alloc(size);
   urosError(chunk == NULL, return NULL,
-            ("Not enough free heap memory to allocate %zu bytes\n", size));
+            ("Not enough free heap memory to allocate %u bytes\n",
+             (unsigned)size));
   return chunk;
 #else
   return uros_lld_alloc(size);
